remove FSL_UID and OEM_UID value for all platforms
FSL UID may be pre-programmed at the time of manufacturing
and it cannot be modified. So t is being removed from the
sample input files.
If someone wants to test with a particular value of these
ID's, then he can modify the same in the respective input
file before generating the header.
OEM UID can also be chosen at OEM's discretion. Removing
both the fields from the demo input files.

Signed-off-by: Aneesh Bansal <aneesh.bansal@freescale.com>
Change-Id: I9e9138eafbcbd019cffe0398d42ad21f1dc17858
Reviewed-on: http://git.am.freescale.net:8181/6900
Reviewed-by: Ruchika Gupta <ruchika.gupta@freescale.com>
Tested-by: Ruchika Gupta <ruchika.gupta@freescale.com>