input files modified for c29x
There were 2 input files present for each ESBC image
(bootscript, rootfs, dtb and uImage)
One was for image present on DDR and second for image on
NOR flash. Since NOR flash is now working on C29x, there
is no need for the images on DDR.
So the corresponding input files are removed.

Signed-off-by: Aneesh Bansal <aneesh.bansal@freescale.com>
Signed-off-by: Gaurav Kumar Rana <gaurav.rana@freescale.com>
Change-Id: I9eeba7591086c56d9a15c0dea9e7658a053e3868
Reviewed-on: http://git.am.freescale.net:8181/7117
Reviewed-by: Ruchika Gupta <ruchika.gupta@freescale.com>
Tested-by: Ruchika Gupta <ruchika.gupta@freescale.com>