LS1043: Input Files added
For LS1043, a single kernel fit image is used to Boot LINUX
So, instead of 3 files for uimage, dtb and rootfs,
we have a single file for signing the kernel.itb image.

The Kernel.itb is to be copied on DDR from Flash via bootscript
So, the address for the image is taken as a DDR address.

Signed-off-by: Aneesh Bansal <aneesh.bansal@freescale.com>
Change-Id: If08188d82f13d73ae68c512f4feaa0ceb3bc9adb
Reviewed-on: http://git.am.freescale.net:8181/39108
Reviewed-by: Ruchika Gupta <ruchika.gupta@freescale.com>
Tested-by: Ruchika Gupta <ruchika.gupta@freescale.com>