LS1021ATWR: Modify tool for ACS write functionality
Previously the acs write function read 64 bytes of image  at a time while
splitting the image for ACS write and the last bytes of the image to be read,
if not equal to 64 bytes, were padded with zero to make it 64 bytes. So,
the tool is modified to write the remaining data, without padding.

Signed-off-by: Vinitha Pillai-B57223 <vinitha.pillai@nxp.com>