Added a new field in input file of pbi creation
For SD Secure boot pbi image is loaded at 8th sector
so a macro APND_IMG_OFF is defined to reduce 8 sectors from
actual offset while appending other images to pbi image. But
in case of NAND, the pbi image is loaded at 0x0 offset,
so the macro is not needed in that case. So, added
a field BOOT_SRC in input field to differentiate b/w boot source

Signed-off-by: Vinitha Pillai-B57223 <vinitha.pillai@nxp.com>