create_header: Add command line options
Following command line options have been added to help
overwrite the values in the input files

--out <file>       Header file name
--in <file>        Input file for signature calculation.
                   This option would override the filename
                   in IMAGE_1 in input_file if present
--app <file>       File to be appended to the header
--app_off <offset> Offset at which file will be appended
                   to the header

There are scenarios when a build script using the tool needs
to modify the input file name or the output header file name.
These command line options provide a way to override the values
as specified in the input file

for eg.

./create_hdr_esbc --in <inp_file>  --out <op file> --app_off <offset> \
                                        --app <file> <input_file>

Signed-off-by: Ruchika Gupta <ruchika.gupta@nxp.com>